Fix slices on reindex APIs
