+++ output/openapi/elasticsearch-serverless-openapi.json
@@ -5471,6 +5471,13 @@
                     "slice_id": {
                       "type": "number"
                     },
+                    "slices": {
+                      "description": "Status of each slice if the delete by query was sliced",
+                      "type": "array",
+                      "items": {
+                        "$ref": "#/components/schemas/_types.ReindexStatus"
+                      }
+                    },
                     "task": {
                       "allOf": [
                         {
@@ -24707,6 +24714,13 @@
                     "slice_id": {
                       "type": "number"
                     },
+                    "slices": {
+                      "description": "Status of each slice if the reindex was sliced",
+                      "type": "array",
+                      "items": {
+                        "$ref": "#/components/schemas/_types.ReindexStatus"
+                      }
+                    },
                     "task": {
                       "allOf": [
                         {
@@ -30725,6 +30739,13 @@
                           "$ref": "#/components/schemas/_types.Retries"
                         }
                       ]
+                    },
+                    "slices": {
+                      "description": "Status of each slice if the update by query was sliced",
+                      "type": "array",
+                      "items": {
+                        "$ref": "#/components/schemas/_types.ReindexStatus"
+                      }
                     },
                     "task": {
                       "allOf": [
@@ -63939,6 +63960,100 @@
         "required": [
           "bulk",
           "search"
+        ]
+      },
+      "_types.ReindexStatus": {
+        "type": "object",
+        "properties": {
+          "slice_id": {
+            "description": "The slice ID",
+            "type": "number"
+          },
+          "batches": {
+            "description": "The number of scroll responses pulled back by the reindex.",
+            "type": "number"
+          },
+          "created": {
+            "description": "The number of documents that were successfully created.",
+            "type": "number"
+          },
+          "deleted": {
+            "description": "The number of documents that were successfully deleted.",
+            "type": "number"
+          },
+          "noops": {
+            "description": "The number of documents that were ignored because the script used for the reindex returned a `noop` value for `ctx.op`.",
+            "type": "number"
+          },
+          "requests_per_second": {
+            "description": "The number of requests per second effectively executed during the reindex.",
+            "type": "number"
+          },
+          "retries": {
+            "description": "The number of retries attempted by reindex. `bulk` is the number of bulk actions retried and `search` is the number of search actions retried.",
+            "allOf": [
+              {
+                "$ref": "#/components/schemas/_types.Retries"
+              }
+            ]
+          },
+          "throttled": {
+            "allOf": [
+              {
+                "$ref": "#/components/schemas/_types.Duration"
+              }
+            ]
+          },
+          "throttled_millis": {
+            "description": "Number of milliseconds the request slept to conform to `requests_per_second`.",
+            "allOf": [
+              {
+                "$ref": "#/components/schemas/_types.DurationValueUnitMillis"
+              }
+            ]
+          },
+          "throttled_until": {
+            "allOf": [
+              {
+                "$ref": "#/components/schemas/_types.Duration"
+              }
+            ]
+          },
+          "throttled_until_millis": {
+            "description": "This field should always be equal to zero in a `_reindex` response.\nIt only has meaning when using the Task API, where it indicates the next time (in milliseconds since epoch) a throttled request will be executed again in order to conform to `requests_per_second`.",
+            "allOf": [
+              {
+                "$ref": "#/components/schemas/_types.DurationValueUnitMillis"
+              }
+            ]
+          },
+          "total": {
+            "description": "The number of documents that were successfully processed.",
+            "type": "number"
+          },
+          "updated": {
+            "description": "The number of documents that were successfully updated, for example, a document with same ID already existed prior to reindex updating it.",
+            "type": "number"
+          },
+          "version_conflicts": {
+            "description": "The number of version conflicts that reindex hits.",
+            "type": "number"
+          },
+          "cancelled": {
+            "description": "The reason for cancellation if the slice was canceled",
+            "type": "string"
+          }
+        },
+        "required": [
+          "batches",
+          "deleted",
+          "noops",
+          "requests_per_second",
+          "retries",
+          "throttled_millis",
+          "throttled_until_millis",
+          "total",
+          "version_conflicts"
         ]
       },
       "_types.TaskId": {
